moved modal into context directory and fixed imports
