+++ react-app/src/Context/index.js
@@ -1,0 +1,3 @@
+import { ModalProvider, Modal, useModal } from "./Modal";
+
+export { ModalProvider, Modal, useModal };
