--- react-app/src/Modal/index.js
@@ -1,3 +1,0 @@
-import { ModalProvider, Modal, useModal } from "./Modal";
-
-export { ModalProvider, Modal, useModal };
