+++ react-app/src/components/Login/LoginModal.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useSelector } from "react-redux";
 
-import { Modal } from "../../Modal";
+import { Modal } from "../../Context";
 import LoginForm from "./LoginForm";
 import styles from "./LoginModal.module.css";
