+++ react-app/src/components/Signup/SignupModal.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { Modal } from "../../Modal";
+import { Modal } from "../../Context";
 import { useSelector } from "react-redux";
 import SignupForm from "./SignupForm";
 import styles from "./SignupModal.module.css";
