+++ react-app/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 
-import { ModalProvider } from "./Modal";
+import { ModalProvider } from "./Context";
 
 import App from "./App";
 import configureStore from "./store";
